Change database for publication
Now expert can add their publication

+++ app/Http/Controllers/ExpertDomainController.php
@@ -7,6 +7,7 @@ use App\Models\ExpertDomain;
 use App\Http\Requests\StoreExpertDomainRequest;
 use App\Http\Requests\UpdateExpertDomainRequest;
 use App\Models\Publication;
+use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Redirect;
 
@@ -41,7 +42,7 @@ class ExpertDomainController extends Controller
         {
             return Redirect::route('login');
         }
-        return view('ManageExpertDomain/listExpertDomain', compact('listexperts'));
+        return view('ManageExpertDomain.listExpertDomain', compact('listexperts'));
     }
 
 
@@ -56,7 +57,7 @@ class ExpertDomainController extends Controller
         {
             return Redirect::route('login');
         }
-        return view('ManageExpertDomain/generateReportExpert', compact('myexperts'));
+        return view('ManageExpertDomain.generateReportExpert', compact('myexperts'));
     }
 
     /**
@@ -64,7 +65,14 @@ class ExpertDomainController extends Controller
      */
     public function create()
     {
-        return view('ManageExpertDomain/addExpertProfile');
+        if(Auth::check())
+        {
+            return view('ManageExpertDomain.addExpertProfile');
+        }
+        else
+        {
+            return Redirect::route('login');
+        }
     }
 
     /**
@@ -72,47 +80,44 @@ class ExpertDomainController extends Controller
      */
     public function store(StoreExpertDomainRequest $request)
     {
-        // dd($request->all());
-        // $request->validate([
-        //     'name' => 'required',
-        //     'emails' => 'required',
-        //     'affiliation' => 'required|max:255',
-        //     'research' => 'required'
-        // ]);
-        
-        // dd($request->name);
-
-        $request->validate([
-            'ed_name' => 'required',
-            'ed_email' => 'required',
-            'ed_phonenum' => 'required',
-            'ed_affiliation' => 'required',
-            'ed_research' => 'required'
-        ]);
-
-        $imageData = null;
-
-        if ($request->hasFile('ed_image'))
+        if(Auth::check())
         {
-            $image = $request->file('ed_image');
-            $imageName = uniqid().'.'.$image->getClientOriginalExtension();
-            $imagePath = $image->storeAs('expert_photos', $imageName, 'public');
+            $request->validate([
+                'ed_name' => 'required',
+                'ed_email' => 'required',
+                'ed_phonenum' => 'required',
+                'ed_affiliation' => 'required',
+                'ed_research' => 'required'
+            ]);
+
+            // $imageData = null;
+
+            if ($request->hasFile('ed_image'))
+            {
+                $image = $request->file('ed_image');
+                $imageName = uniqid().'.'.$image->getClientOriginalExtension();
+                $imagePath = $image->storeAs('expert_photos', $imageName, 'public');
+            }
+            else
+            {
+                $imagePath = "null";
+            }
+
+            $ed = new ExpertDomain;
+            $ed->expert_domain_names = $request->ed_name;
+            $ed->expert_domain_emails = $request->ed_email;
+            $ed->expert_domain_phonenumbers = $request->ed_phonenum;
+            $ed->expert_domain_affiliation = $request->ed_affiliation;
+            $ed->expert_domain_research_title = $request->ed_research;
+            $ed->expert_domain_designation = implode(',', (array) $request->ed_designation);
+            $ed->expert_domain_image = $imagePath;
+            $ed->platinum_id = Auth::user()->getPlatinum()->id;
+            $ed->save();
         }
         else
-        {
-            $imagePath = "null";
+        { 
+            return Redirect::route('login');
         }
-
-        $ed = new ExpertDomain;
-        $ed->expert_domain_names = $request->ed_name;
-        $ed->expert_domain_emails = $request->ed_email;
-        $ed->expert_domain_phonenumbers = $request->ed_phonenum;
-        $ed->expert_domain_affiliation = $request->ed_affiliation;
-        $ed->expert_domain_research_title = $request->ed_research;
-        $ed->expert_domain_designation = implode(',', (array) $request->ed_designation);
-        $ed->expert_domain_image = $imagePath;
-        $ed->platinum_id = Auth::user()->getPlatinum()->id;
-        $ed->save();
 
         return Redirect::route('my-expert');
     }
@@ -135,7 +140,7 @@ class ExpertDomainController extends Controller
         {
             return Redirect::route('login');
         }
-        return view('ManageExpertDomain/viewExpertProfile', compact('expert'), compact('publications'));
+        return view('ManageExpertDomain.viewExpertProfile', compact('expert'), compact('publications'));
     }
 
     /**
@@ -151,7 +156,7 @@ class ExpertDomainController extends Controller
         {
             return Redirect::route('login');
         }
-        return view('ManageExpertDomain/editExpertProfile', compact('expert'));
+        return view('ManageExpertDomain.editExpertProfile', compact('expert'));
     }
 
     /**
@@ -202,22 +207,59 @@ class ExpertDomainController extends Controller
      */
     public function destroy(ExpertDomain $expertDomain, string $id)
     {
-        $ed = $expertDomain->findorFail($id);
-        $ed->delete();
-
+        if(Auth::check())
+        {
+            $ed = $expertDomain->findorFail($id);
+            $ed->delete();
+        }
+        else
+        {
+            return Redirect::route('login');
+        }
         return Redirect::route('my-expert');
     }
 
-    // public function addpublication(string $id)
-    // {
-    //     return view('ManageExpertDomain/uploadExpertPublication', ["id" => $id]);
-    // }
+    public function addpublication(string $id)
+    {
+        return view('ManageExpertDomain/uploadExpertPublication', ["id" => $id]);
+    }
 
-    // public function publication(Request $request ,string $id)
-    // {
-    //      $result = new PublicationController;
-    //      $result->store($request);
+    public function publication(Request $request ,string $id)
+    {
+        if(Auth::check())
+        {
+            $request->validate([
+                'P_authors' => 'required|string',
+                'P_title' => 'required|string',
+                'P_published_date' => 'required|date',
+                'P_type' => 'required|string', // Ensure this matches your input type
+                'P_volume' => 'required|integer',
+                'P_issues' => 'required|integer',
+                'P_pages' => 'required|integer',
+                'P_publisher' => 'required|string',
+                'P_description' => 'required|string',
+                'P_path' => 'required|string',
+            ]);
 
-    //     return Redirect::route('view-expert.id', $id);
-    // }
+            $publication = new Publication;
+            $publication->expert_domain_id = $id;
+            $publication->platinum_id = NULL;
+            $publication->P_authors = $request['P_authors'];
+            $publication->P_title = $request['P_title'];
+            $publication->P_published_date = $request['P_published_date'];
+            $publication->P_type = $request['P_type'];
+            $publication->P_volume = $request['P_volume'];
+            $publication->P_issues = $request['P_issues'];
+            $publication->P_pages = $request['P_pages'];
+            $publication->P_publisher = $request['P_publisher'];
+            $publication->P_description = $request['P_description'];
+            $publication->P_path = $request['P_path'];
+            $publication->save();
+        }
+        else
+        {
+            return Redirect::route('login');
+        }
+        return Redirect::route('view-expert.id', $id);
+    }
 }
